Move multilayergraph methods  to abstractmultilayerugraph
