+++ src/multilayergraph.jl
@@ -9,7 +9,7 @@ A concrete type that can represent a general multilayer graph.
 - `layers::OrderedDict{ Tuple{Int64,Int64}, Layer{T,U,G}}`: the ordered dictionary containing all the layers of the multilayer graph. Their underlying graphs must be all undirected.
 - `interlayers::OrderedDict{ Tuple{Int64,Int64}, Interlayer{T,U}}`: the ordered dictionary containing all the interlayers of the multilayer graph. Their underlying graphs must be all undirected.
 """
-mutable struct MultilayerGraph{T,U} <: AbstractMultilayerGraph{T,U}
+mutable struct MultilayerGraph{T,U} <: AbstractMultilayerUGraph{T,U}
     adjacency_tensor::Array{U,4}
     layers::OrderedDict{Tuple{Int64,Int64},Layer{T}}
     interlayers::OrderedDict{Tuple{Int64,Int64},Interlayer{T}}
@@ -125,25 +125,7 @@ function MultilayerGraph(
     return multilayergraph
 end
 
-"""
-    add_layer!(mg::M,layer::L; interlayers_type = "multiplex") where { T, U, G<: AbstractGraph{T}, M <: MultilayerGraph{T, U}, L <: Layer{T,U,G}}
 
-Add layer `layer` to `mg`. Also add `Interlayer`s of type `interlayers_type` (can only be `"multiplex"`) between the new layer and all the other ones. 
-"""
-function add_layer!(
-    mg::M, new_layer::L; new_default_interlayers_type="multiplex"
-) where {T,U,G<:AbstractGraph{T},M<:MultilayerGraph{T,U},L<:Layer{T,U,G}}
-    # Check that the layer is directed
-    !istrait(IsDirected{typeof(new_layer.graph)}) || throw(
-        ErrorException(
-            "The `new_layer`'s underlying graph $(new_layer.graph) is directed, so it is not compatible with a `MultilayerGraph`.",
-        ),
-    )
-
-    if new_default_interlayers_type == "multiplex"
-        _add_layer!(mg, new_layer; new_default_interlayers_type=SimpleGraph{T})
-    end
-end
 
 """
     specify_interlayer!(mg::M, new_interlayer::In; symmetric_interlayer_name::String) where { T, U, G<: AbstractGraph{T}, M <: MultilayerGraph{T, U}, In <: Interlayer{T,U,G}}
@@ -183,7 +165,7 @@ function specify_interlayer!(
     symmetric_interlayer_name::String="interlayer_$(layer_2)_$(layer_1)",
     forbidden_vertices::Tuple{Vararg{MultilayerVertex{T}}},
     forbidden_edges::Tuple{Vararg{NTuple{2,MultilayerVertex{T}}}},
-) where {T,U,G<:AbstractGraph{T},M<:MultilayerGraph{T,U}} # 
+) where {T,U,G<:AbstractGraph{T},M<:MultilayerGraph{T,U}}
     !istrait(isDirected{G}) || throw(
         ErrorException(
             "The new interlayer's underlying graphs $(graph) is directed, so it is not compatible with a `MultilayerGraph`.",
@@ -204,12 +186,6 @@ end
 
 # Graphs.jl's internals extra overrides
 
-function Graphs.degree(
-    mg::M, v::V
-) where {T,M<:MultilayerGraph{T,<:Real},V<:MultilayerVertex{T}}
-    return indegree(mg, v)
-end
-
 """
     add_edge!(mg::M, src::V, dst::V, weight::U) where { T, U <: Real, M <: MultilayerGraph{T,U}, V <: MultilayerVertex{T}}
 
@@ -218,6 +194,7 @@ Add edge from `src` to `dst` with weight `weight` to `mg`.
 function Graphs.add_edge!(
     mg::M, src::V, dst::V, weight::U
 ) where {T,U<:Real,M<:MultilayerGraph{T,U},V<:MultilayerVertex{T}}
+
     if !(has_vertex(mg, src) && has_vertex(mg, dst))
         return false
     end
@@ -346,27 +323,7 @@ function Graphs.add_edge!(
     return added
 end
 
-"""
-    add_edge!(mg::M, me::E) where { T, U <: Real, M <: MultilayerGraph{T,U}, E <: MultilayerEdge{MultilayerVertex{T},U}}
 
-Add weighted edge `me` to `mg`.
-"""
-function Graphs.add_edge!(
-    mg::M, me::E
-) where {T,U<:Real,M<:MultilayerGraph{T,U},E<:MultilayerEdge{MultilayerVertex{T},U}}
-    return add_edge!(mg, src(me), dst(me), weight(me))
-end
-
-"""
-    add_edge!(mg::M, me::E) where { T, U, M <: MultilayerGraph{T,U}, E <: MultilayerEdge{MultilayerVertex{T},Nothing}}
-
-Add unweighted edge `me` to `mg`.
-"""
-function Graphs.add_edge!(
-    mg::M, me::E
-) where {T,U,M<:MultilayerGraph{T,U},E<:MultilayerEdge{MultilayerVertex{T},Nothing}}
-    return add_edge!(mg, src(me), dst(me))
-end
 
 """
     rem_edge!(mg::M, V1::V, V2::V) where { T, U, M <: MultilayerGraph{T,U}, V <: MultilayerVertex{T}}
@@ -433,63 +390,8 @@ function Graphs.rem_edge!(
     end
 end
 
-"""
-    is_directed(m::M) where { M <: MultilayerGraph}
 
-Returns `true` if `m` is directed, `false` otherwise. 
-"""
-Graphs.is_directed(m::M) where {M<:MultilayerGraph} = false
-
-"""
-    is_directed(m::M) where { M <: Type{ <: MultilayerGraph}}
-
-Returns `true` if `m` is directed, `false` otherwise. 
-"""
-Graphs.is_directed(m::M) where {M<:Type{<:MultilayerGraph}} = false
-
-# TODO:
-# it may be not well inferred
-# function get_projected_monoplex_graph end #approach taken from https://github.com/JuliaGraphs/Graphs.jl/blob/7152d540631219fd51c43ab761ec96f12c27680e/src/core.jl#L124
-"""
-    get_projected_monoplex_graph(mg::M) where {M<: MultilayerGraph}
-
-Get projected monoplex graph (i.e. that graph that as the same nodes as `mg` but the link between node `i` and `j` has weight equal to the sum of all edges weights between the various vertices representing `i` and `j` in `mg`, accounting for both layers and interlayers). See [De Domenico et al. (2013)](https://doi.org/10.1103/PhysRevX.3.041022).
-"""
-function get_projected_monoplex_graph(mg::M) where {M<:MultilayerGraph}
-    projected_monoplex_adjacency_matrix = dropdims(
-        sum(mg.adjacency_tensor; dims=(3, 4)); dims=(3, 4)
-    )
-
-    isapproxsymmetric(projected_monoplex_adjacency_matrix) ||
-        throw(ErrorException("Adjacency / distance matrices must be symmetric"))
-
-    symmetric_projected_monoplex_adjacency_matrix = Symmetric(
-        projected_monoplex_adjacency_matrix
-    )
-
-    return SimpleWeightedGraph{M.parameters[1],M.parameters[2]}(
-        symmetric_projected_monoplex_adjacency_matrix
-    )
-end
-
-# function get_overlay_monoplex_graph end #approach taken from https://github.com/JuliaGraphs/Graphs.jl/blob/7152d540631219fd51c43ab761ec96f12c27680e/src/core.jl#L124
-"""
-    get_overlay_monoplex_graph(mg::M) where {M<: MultilayerGraph}
-
-
-Get overlay monoplex graph (i.e. the graph that has the same nodes as `mg` but the link between node `i` and `j` has weight equal to the sum of all edges weights between the various vertices representing `i` and `j` in `mg`, accounting for both layers and interlayers). See [De Domenico et al. (2013)](https://doi.org/10.1103/PhysRevX.3.041022).
-"""
-function get_overlay_monoplex_graph(mg::M) where {M<:MultilayerGraph}
-    projected_overlay_adjacency_matrix = sum([
-        mg.adjacency_tensor[:, :, i, i] for i in 1:size(mg.adjacency_tensor, 3)
-    ])
-    return SimpleWeightedGraph{M.parameters[1],M.parameters[2]}(
-        projected_overlay_adjacency_matrix
-    )
-end
-
-#TODO:
-# We should decide whether to implement is_directed() for the MultilayerGraph. If we do implement it, the normalization factor (2 or 1) will be decided at the MultilayerGraph level,else at the 
+# Multilayer-specific functions
 # function get_graph_of_layers end #approach taken from https://github.com/JuliaGraphs/Graphs.jl/blob/7152d540631219fd51c43ab761ec96f12c27680e/src/core.jl#L124
 """
     get_graph_of_layers(mg::M) where {M <: MultilayerGraph}
@@ -522,48 +424,4 @@ function get_graph_of_layers(
     )
 end
 
-"""
-    von_neumann_entropy(mg::M) where {T,U,  M <: MultilayerGraph{T, U}}
 
-Compute the Von Neumann entropy of `mg`, according to [De Domenico et al. (2013)](https://doi.org/10.1103/PhysRevX.3.041022). Only for undirected multilayer graphs.
-"""
-function von_neumann_entropy(mg::M) where {T,U,M<:MultilayerGraph{T,U}}
-    num_nodes = length(nodes(mg))
-    num_layers = length(mg.layers)
-
-    # multistrength tensor
-    Δ = ein"ijkm,ik,nj,om -> njom"(
-        mg.adjacency_tensor,
-        ones(T, size(mg.adjacency_tensor)[[1, 3]]),
-        δ(num_nodes),
-        δ(num_layers),
-    )
-
-    # trace of Δ
-    tr_Δ = ein"iikk->"(Δ)[]
-
-    # multilayer laplacian tensor
-    L = Δ .- mg.adjacency_tensor
-
-    # multilayer density tensor
-    ρ = (1.0 / tr_Δ) .* L
-
-    eigvals, eigvects = tensoreig(ρ, [2, 4], [1, 3])
-
-    #=     # Check that we are calculating the right eigenvalues
-        lhs = ein"ijkm,ik -> jm"(ρ,eigvects[:,:,1])
-        rhs = eigvals[1].*eigvects[:,:,1]
-        @assert all(lhs .≈ rhs)
-        # Indeed we are =#
-
-    Λ = get_diagonal_adjacency_tensor(eigvals, size(mg.adjacency_tensor))
-
-    # Correct for machine precision
-    Λ[Λ .< eps()] .= 0.0
-
-    log2Λ = log2.(Λ)
-
-    log2Λ[isinf.(log2Λ)] .= 0
-
-    return -ein"ijkm,jimk ->"(Λ, log2Λ)[]
-end
